requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
numpy==1.23.3
Pillow==9.2.0
tflite-runtime==2.10.0